refactor & extend
- refactored fixtures into reusable functions
- refactored test files by contract name
- renamed them all accordingly
- created a test file for the Tools contract and begun layout out tests for calcLiquidityUnits()
- refactored out the test variables to their own file (this will keep them consistent throughout tests now that they are split up)
- adjusted some notes throughout
